Store explicit valid state for node_candidates
