+++ include/osr/ways.h
@@ -27,6 +27,7 @@
 #include "utl/verify.h"
 #include "utl/zip.h"
 
+#include "osr/extract/tags.h"
 #include "osr/point.h"
 #include "osr/routing/turns.h"
 #include "osr/types.h"
@@ -389,6 +390,19 @@ struct ways {
   mm_vec<pair<way_idx_t, string_idx_t>> way_conditional_access_no_;
 
   multi_counter node_way_counter_;
+};
+
+struct way_extra_properties {
+  // Properties only required for extract
+  explicit way_extra_properties(tags const&);
+
+  constexpr bool is_foot_usable() const { return is_foot_usable_; }
+  constexpr bool is_car_usable() const { return is_car_usable_; }
+  constexpr bool is_parking_aisle() const { return is_parking_aisle_; }
+
+  std::uint8_t is_foot_usable_ : 1 = 0U;
+  std::uint8_t is_car_usable_ : 1 = 0U;
+  std::uint8_t is_parking_aisle_ : 1 = 0U;
 };
 
 }  // namespace osr
